Device allocator should not use reserved dev names.
According to
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/device_naming.html,
on HVM instances, sda1 and xvda are reserved for root partition. Also,
the available block device range for xvd follow "/dev/xvd[b-c][a-z]"
pattern.

However commit 1ace946 violates the reserved device names and the naming
logic. This causes volume attachment issues. For example, for instances
where /dev/xvda is root partition, the driver tries to create a new
device with the same reserved name.

This change fixes that.

Issue #292

Signed-off-by: Anusha Ragunathan <anusha.ragunathan@docker.com>