Add controller support for custom EC2 Endpoints
Expose the AWS_EC2_ENDPOINT environment variable as a ConfigMap for the Controller. This allows deployment of the controller to environments using EC2-compatible clouds such as Eucalyptus/CloudStack or Government Community clouds disconnected from the internet. The controller already has support for the environment variable, it just needs to be exposed in the config map.

This was tested on Eucalyptus and confirmed to be functioning properly.