[helm] Add support for managing StorageClass
So after installing Helm chart, there is no need to manually create a
StorageClass object. Usually deploying CSI driver and storage class is
administrator role and consuming it is up to the user, so I think
shifting this responsibility to Helm chart make sense.

Right now this is implemented without support for any parameters. I
believe they can be added on demand.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>