Aws client config: increase MaxRetries
This commit increase the number of retries to a high number (8) so
that the AWS framework retries failing requests until the request
context times-out.

By default, when throttled, the aws client retries a given request
3 times with exponentially increasing delay between retries.
The default number of retries is still low enough that requests are
only going to be retried for a few seconds.
Instead of retrying for a few seconds, we should retry until the
request context times out.

This should reduce the occurence of #754