Implement automatic IOPS and Throughput reconciliation for gp3 volumes
Using gp3 StorageClasses in highly volatile environments, where for some
reason there's a necessity of resizing volumes frequently, makes it hard
not to have a way to reconcile IOPS and Throughput automatically.

The simplest approach is to mimic the gp2 behavior for gp3 volumes,
calculating the value for IOPS and Throughput programmatically and
updating everything all at once.