Stop treating prefixes as magic in DeviceManager
DM treats the prefixes "/dev/xvd" and "/dev/sd" as magic. This leads to
a bug where volumes with a prefix starting with "/dev/sd" (for example,
"/dev/sdf") will get incorrectly mangled into a "/dev/xvd" form (i.e.
converting "/dev/sdf" to "/dev/xvdf"). This will cause a volume that is
already mounted and prefixed with "/dev/sd" (for example, a volume
manually mounted via the AWS console) to get stuck in an infinite loop
as the node will believe it is mounted at the wrong path.

With this change, instead of storing only the end of the path in the
inflight and existing volumes map, we instead store the entire device
path. The previously leaky abstraction from the allocator to store only
the end of the path is contained to the allocator, which has a new
parameter for the desired prefix.

This commit also vastly increases the range of paths the allocator can
generate from 52 simultaneously mounted volumes to 650.

Signed-off-by: Connor Catlett <conncatl@amazon.com>