Grab snow device path via last character rather than trimming
Due to #1518, volumes now start with /dev/xvda rather than /dev/xvdb.
Because of this, the hardcoded path in the snow code stopped working,
breaking mounting on snow devices. This change instead creates the
snow device path based on the last character of the generated path,
rather than trying to do it via trimming, which is unreliable.

Snow devices are currently limited to 10 volumes attached which makes
this 100% safe, but this will need to be revisited if snow ever adds
suport for >26 volumes.

Signed-off-by: Connor Catlett <conncatl@amazon.com>