chore: default to policy/v1
policy/v1beta1 was removed almost a year ago in kubernetes 1.25. The default for the `PodDisruptionBudget` should be modernized to `policy/v1` and fallback to `v1beta1` if `policy/v1` is not supported in the cluster. The `.Capabilities` object does not validate cluster capabilities when generating templates so is an imperfect solution for ensuring we're using the right api version.

Signed-off-by: Noah Krause <krausenoah@gmail.com>