fix: just check for the base api version.
`helm template` uses the `chartutil.DefaultCapabilities` in the `.Capabilities.APIVersions` check. This is limited to the Kubernetes group/version and omits the kind. For the purpose of making the sure the PDB is being installed with the correct policy version, just checking if `policy/v1` exists in the cluster should be enough.

Addresses: https://github.com/helm/helm/issues/11809
Signed-off-by: Noah Krause <krausenoah@gmail.com>