Add otel gRPC tracing support
The feature is disabled by default and hidden behind a the flag
--enable-otel-tracing. When enabled, all the gRPC calls made by the
driver will be instrumented and can be forwarded to an opentelemetry-
compatible collector.
The configuration of opentelemetry tracing can be done in the chart with
values `controller.otelTracing` and `node.otelTracing`. The service name
and the endpoint that will be used by the sdk can be configured with
`otelServiceName` and `otelExporterEndpoint` for both components.