Add option to disable discovery of attached volumes from instance metadata
Add --disable-volume-attach-limit-from-metadata to disable counting of
attached volumes from instance metadata.

Instance metadata contains block volumes that were attached at the time the
instance was started and then it's never updated. In case a node was
rebooted with CSI volumes attached, the metadata includes these CSI volumes
and the CSI driver "reserves" an attachment slot for them forever, even
when they are detached later.

The option disables counting of attached volumes when computing the attach
limit, because it's not able to distinguish a CSI volume (should not have a
reserved attachment, because scheduler will count it on its own) and a root
disk / volume attached by the system admin (should have a reserved
attachment).