fix(monitoring): use "port" in service monitor
The ServiceMonitor has two ways to specify the port of the metrics
endpoint. `port` is used for the port name of the Service, `targetPort`
is used for the port number on the Pod. A port used as `targetPort` also
has to be specified in the port list of the Pod resource. Since this is
not the case for ebs-csi-node and ebs-csi-controller, `port` has to be
used.

Signed-off-by: Timon Engelke <timon.engelke@inovex.de>